Add form to see save places; Refactor architecture
