+++ src/views/components/formSavedPlaces.js
@@ -1,0 +1,7 @@
+`<div class="black-bg hidden" >
+    <div class="add-city-form" >
+        <h4>Your Cities</h4>
+        <div class="your-cities"></div>
+        <input type="text" placeholder="Add a new city">    
+    </div>
+</div>`
