+++ src/views/components/weatherPlaces.js
@@ -1,0 +1,7 @@
+`
+<div id="carrousel-places">
+    <div class="weather-img-cont">
+        <div class="img-place" id="img-1"></div>
+        <p>{cityWeather.city.country}, {cityWeather.city.name}</p>
+    </div>
+</div>`
