+++ src/views/components/weatherStats.js
@@ -1,0 +1,20 @@
+export default todayWeather = (we) => {  
+    const weatherMenu =    `<div id="time-weather">
+                        <button id="week">week</button>
+                        <button id="month">month</button>
+                        <button id="3-month">3-month</button>
+                        <button id="5-month">5-month</button>
+                    </div> `
+            
+    const weatherStats = `<div class="weather-stats">
+                                <div>sunday</div>
+                                <div class="humidity"><i></i><p>54%</p></div>
+                                <div>O</div>
+                                <div>23 C°</div>
+                                <div class="line">
+                                    <div class="left-probable"></div>
+                                    <div class="right-probable"></div>
+                                </div>
+                                <div>55C°</div>
+                          </div>`
+}
