+++ webpack.config.js
@@ -2,7 +2,7 @@ const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
-    entry: './src/controllers/weatherAppController.js',
+    entry: ['./src/controllers/weatherAppController.js', './src/db/apiCalls.js'],
     output: {
         filename: 'main.js',
         path: path.resolve(__dirname, 'dist'),
